requirement.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastmcp==2.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mcp==1.7.1
mdurl==0.1.2
mysql-connector-python==9.3.0
openai==1.77.0
openapi-pydantic==0.5.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyodbc==5.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette>=0.41.3
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
websockets==15.0.1
fastapi>=0.110.0
jinja2==3.1.3
